requirements.txt
flask
flask-cors
openai
pydantic
python-dotenv
